Added TreeNode class and parse tree parser.
All classes which can be defined via cmd line do now register themselves via some names (
which shall be used in the command line to define them) in a register of their subpackage
(e.g. all instanctiable conditions are registered in a register of the conditions subpackage.

TreeNode class added with large functionalities.

Parser added which transforms a command line input (for an object to create which can contain
other object definitions. This does not parse the whole command line input, only for a single
object!) into a parse tree. Each class has then to be able to use this parse tree to create an
object of itself.