Added framework for parsing from command line the objects.
The given parse tree is now translated to a structure of nested objects. Via class ClassArguments
the arguments a class constructor has can be described and thus, we can automatically extract
the arguments, if needed also parse them or convert them, and give them to the object constructor.
this is implemented for variable now and the base/advanced conditions. They are also tested.