Improved Parsing Framework alot. Tested Variable and condition subpackage. Made registers in the packages to global register class. The Register (for looking up class references) and the caching (for defining variables and reuse them via name or id) work now on arbitrary class granularity. that means for register: every class is registered for itself and its base classes. If an object under construction expects an parameter of class X, then it can look up for X the list of all valid classes (X and the children of X).
The caching caches every object now under its class and its base classes. Again,
if parameter needs class  X, we can now look up for X and for objects of
subclasses.

The result is that only classes without common child/ancestor class can reuse
ids for caching (the first ancestor 'object' class is ignored.)

The ClassArgument class now supports that the type of a parameter can be defined
as a list of possible types (thus, the values for a parameter must not be anymore
in an inheritance relation).