Added Bridges, Environments, Task
Bridges define some operations (The base bridge shall be subclassed per
different setting where to use (e.g. in Sampler) and those subclasses
shall ebe subclassed with the different possible functionalities the
bridge shall provide. The idea is that We do not create an abundant
number of subclasses, but can decide for some objects (e.g. sampler)
how they do something via giving them an object which does the
operation.

Task objects shall encapsulate expensive task. They are submitted to an
Environment.

Environments manage how to run Task objects. E.g. a Slurm environment
submits the task to the queue and waits for their completion. Depending
on the parameters of the environment tasks can be run in parallel.