CMAKE scripts with packages & Neural Network Heuristic
CMAKE build configuration extended to define packages for plugins. Those packages
are loaded. If the package is not one of the defaults known by CMAKE, then the user
can provide their own FindPACKAGE script for loading.
This is used to load the dependencies for Neural Networks. In the example case:
Tensorflow, Protobuf, Eigen.
The script looks in the environment variable PATH_PACKAGE for path hints.

SOME HACKS ARE STILL IN THERE TO SOLVE


secondly, the base infrastructure for networks in FD were added and an example network
(syimple protobuf network from my masterthesis). This still has to be debugged.