Reviewed CMake configuration, debugged Network Code
The cmake configuration is now reviewed. An additional
flag in the packages now tells if the include_directories
shall ignore warnings (e.g. for dependencies out of our
reach).
Currently I (Patrick) need to compile as dynamic build,
because Tensorflow could only be build as dynamic library.

The code for the networks in Fast-Downward works now, too.
A simple heuristic is implemented which can use any
AbstractNetwork which is a heuristic. A simple network
class using tensorflow (and protobuf for storing) is also
implemented.

PROBLEM: The argument parsing is currently in Fast-Downward
converting everything to lower case, this is problematic
with string which are case-sensitive (e.g. file paths).
I am working on this.