Py2 Compatibility | Sampling Script | Translate Package
The static methods in training were annotated with @staticmethod

The translate package was modified that it can now be imported as a python package,
but still can be called via the translate.py script which is within the package.

To sample data with Fast-Downward the script fast-sample.py can now be used.
The training package is not finished, thus this script eases access to some functionalities
of the package. (FastDownwardSamplerBridge and DirectorySampler)
1. It allows to sample for some problems data
2. It allows to traverse in a directory and sample for the problems found data.
Both modes have a lot of options, consult the help menu for more information.
The data sampled by FD can also be converted into the following formats:
    FD = PDDL dump format of FD
    FDFull = For all var-value pairs of FD its PDDL name and + if True, - if False
    FDFullShort = Like before, but pairs are sorted alphabetically and the names not printed
    Full = ALL PDDL predicates of the  task with + and - if true resp. False
    FullShort = For all predicates (in alphabetical order) + resp - if they are true resp. false
    Objects = definition of objects in the task and the predicates true in the state