Python2 Compatibility | fast-sample.py Features| slurm-sample.sh
The declaration of abstract classes was reworked to support python2
without additional dependencies (previously required future.utils)

Fast-sample.py --traverse got a --batch option. Previously all detected
problems in a traversal were given to a single invocation of the named
script (if a specific script was named via --execute). Now at most batch
many problems are given to a single invocation and more invocations are
done (this can be useful if working on a cluster and every invocation
submits a new job)
Additionally, the FastDownwardSamplingBridge got a pruning option which
can be used from fast-sample.py via --prune. Duplicate entries within
the SAME sampling search run are removed.

The slurm-sample.sh script (which submits an invocation of fast-sample.py
to a slurm server) was reworked The TARGET_DIR variable which specifies
where the sampled data shall be stored is not anymore the first argument
provided to the script (because slurm does not support commandline
arguments at this position), but an environment variable which can
be set in sbatch via --export="VAR=XYZ". This variable is used to
define --target-folder in the fast-sample.py script which it starts.
TARGET_DIR is now optional. If not given, it will not be provided to
fast-sample.py