Data Container | Sample Passing to Invoker
Created container for managing sampled data. Data is now stored in this container
in batches. One batch collects all samples where the data field of the sample
have the same size (needed for Tensorflow and Theano).

Sampled data is now not only stored, but also passed to the invocing sampler and
from there to the thing invocing the sampler (e.g. a training schema). Additonally,
The Fast-Downward sampler has learned to load data.