First Version of Network Features (src.training.network)
# fast-training.py

Simple via command line configurable script to perform training of networks.
First define your network via a string (see src.training.parser/parser_tools
and some network how an object is build from a string). Then root directories
follow in which data is looked for. Add any of the other parameters as desired.

The script constructs an network, loads the data in the defined directories via a
DirectorySampler object, then trains the network with the data and evaluates the
network.

If test and trainings data is given, all elements from the test data are pruned from
the training data. If desired and the network supports it a DomainProperties
object about the problems at hand is provided to the network.


# Training Modules (esp. Networks)

## Sampling
- Sampling does now return the data in a SizeBatchData object. And if sampling was
called with a SizeBatchData object, it adds its new samples to that object.
- new Sample Bridge for loading data (LoadSampleBridge)
- duplicate pruning on loaded samples
- Magic word at top of data s.t. the opening function can be automatically determined.
  Additionally, test reading functions till found correct function (currently:
  plain text and gzip).


## Networks

Networks were reworked, extended, and some concrete implementation given.

- New variable for default output directory.

- ABC methods to get supported formats of concrete networks (default network format,
storage formats, formats for loading, formats in which given states can be)

- split initialization into general initialization and model initialization. if
  loading, the model initialization is replaced by _load.

- Load method

### Keras networks

Basic keras network. General training settings, storing, loading, training, evaluating,
analysing, supported formats

KerasDomainPropertiesNetwork: Keras network that accepts a DomainProperties object to
prune static atoms in its input conversion.

Concrete Network: KerasDynamicMLP(KerasDomainPropertiesNetwork)

Simple MLP Network using Keras. The number of hidden layers and output units can be
defined, as well as some basic training settings.

## Data

###SizeBatchData

Stores samples with a fixed number of fields (e.g current state,
goal state, heuristic value). It can store samples for different types of things (e.g.
optimal traces, satisficing traces), by default everything is stored under the type None.
The data is organized in groups. Within a group all fields have the same size (e.g if
the first field current state is a list of pddl predicates, then for different problems
this list can be differently long. Within group1 are only states where all fields have
the same length, that includes that all states in field 1 have the same number of atoms).

- A pruning function can be provided which shall compute "unique" hash values for each
sample used for pruning.

- all fields can have descriptions and a map can store meta description for the whole
object.

Given another SizeBatchData object, all samples of the other object appearing in our
object can be deleted via 'remove_duplicates_from'.

###SampleBatchData

Specialization of SizeBatchData for samples I use. it has five special attributes:

field_current_state, field_goal_state, field_other_state, field_action, field_heuristic

which tell in which field to find the respective information.

## Similarities

A module for defining similarities on samples was added. Similarities can now also be parsed
from the command line. A general framework is written. Currently only the hamming distance
between two lists is implemented as distance.


## KerasDataGenerator
Keras Generator which yields batches of data for training and testing. the data within a
batch has always the same size (in all its dimensions), the data in different batches
has different sizes.

Takes SampleSizeBatch list as input data, as well as, descriptions of the fields to return
as X and for Y data. Some converters to convert the extracted batch (which is in the format
of the input SampleSizeBatches) to the required values, and more.

## Domain Properties

DomainProperties object analyses a domain and the given problems and stores properties of it.
Currently only fixed universes (sets of problems where each problem has the same objects) can
be analysed.

If detects if the world is a FIXED UNIVERSE and if provide the list of OBJECTS. Which TYPES are
instanciated in the problems (e.g. some types in the hierarchy for grouping could never
instanciated objects of their type, but only of subtypes). Atoms STATIC for all problems,
and FLEXIBLE Atoms for at least one problem (the union of them is the set of all atoms) and STATIC
FLEXIBLE atoms per problem. Using this information if determines which PREDICATES have STATIC
instantiation schemas (e.g. road(l1, l1) is always static) or even where all groundings of
a predicate are static (does this occur?)

## Reworks
- Meta tag extracting, modification made more flexible

## Canonicalization
Functions which canonicalize Atoms. at(SB - Location, MyTruck - Truck) will be canonicalized to at(Location-1, Truck-1). Different input formats and output formats are available (e.g. string or atom). The canonical form is always the type which fits in there plus the minimum possible index. Multiple atoms can
be together given for canonicalization.

In the case where multiple atom fields share a type, different objects or the same object could be used
in there (e.g. road(Location, Location)), then all possible canonicalizations are provided
(e.g. road(Location-1, Location-1), road(Location-1, Location-2)).

And some more functionalities related to this.


# FD StateGoalNetwork
This network takes as input a state (as pddl) and the goals of the problem
(described with the same pddl atoms where 1 if this atom is part of the
goal, otherwise 0). Any number of outputs can be produced.
Additional input parameters: List of atoms of the state too feed and their
default values (e.g. if some atoms are pruned by the translator) Only the
atoms in the list given are set by Fast Downward (thus, some atoms can
be ignored, if they are not in the list)


# Classes

##
## Example of StateGoalNetwork

Example can be found in /eamples/state_goal_network. It contains a trained
network two pddl problems and a run file (which is currently only running)
one of the pddl files.

# Reworks

## Misc
- better  python2/3 compatibility
    - with_metaclass replaced with other techniques not needed external dependencies

## Fast Downward ProtobufNetwork & Problem Network
Protobuf Network has now only the input vector<pair<name,tensor>>, names of
the output layers to fetch, and a for the output vector<Tensor>. Previously,
the input/output layer names where both managed by ProtobufNetwork, this
has now to be done by the concrete networks. This allows them to define
specific (different) parameters to parse those names (especially relevant,
in the case of multiple input or output layers). Thus, ProblemNetwork
got code for managing the input_layer and output_layer names.