Dependency Management System
Dependency flags can be defined in dependencies.py. Modules can decide to
run code differently/not all all depending on those flags. E.g. the flag
can prevent the loading of some sub modules or of some external libraries.

To set the dependencies for a script without knownig or changing the
dependencies defined in dependencies.py, the modoule has to be loaded
as standalone. Example code for this can be found in the __doc__ of
dependencies.py.