Keras Callback functions
Added Callback function support to the KerasNetwork. A Base class
was created to enhance the callbacks by features we need and to be
parsable from cmd line.

A KerasCallbackWrapper class can be used to easily wrap an existing
Keras Callback class into the framework. The ModelCheckpoint callback
was wrapped in here.

KerasProgressChecking is a new Callback which checks it the network
was able to learn something in the first n epochs.