Uniform Sampling & Framework Got OutputStreams & Rework Sampling
-FAST DOWNWARD:
SamplingTechniques have now an optional list parameter for evaluators.
For every sampled task, they check if the task is solvable. If not,
a new task is sampled instead.

Added uniform_none sampling technique to fast downward. This chooses random
initial stattes and tries to find plans to the original goal.


-TRAINING FRAMEWORK:
At some point it might happen (e.g. now for me) that you sample only
few samples per problem (extreme case only one) and you do not want
that every sample is in its own file. Thus, sampling shall append those,
but at the same time, it shall append only to the samples of the current
run not of the previous sampling run. Additionally, you might want to be
a bit more precise, the samples from those problems shall be merged into
this file, the samples from these problems into another file (e.g. an
totally independent test set). The waiting is over.


--SizeBatchData:
Allows now splitting of fraction of itself (but not more than 100%)
Allows now copying itself (possible deep copying) or simply copying
    its structure without data
Allows adding all entries of another (structural fitting) SizeBatchData
    object


--MatchRegex
A simple Regex object which can be parsed by the framework from the
commandline and has a check(arg) function

--StreamDefinition
Base for defining output streams.
output: Tells where to store the output either
    via telling the file, telling the directory (then it stored file asdf/.../f.g
    in directory/f.g) or nothing (then stored in asdf/../f.suffix with suffix
    defined).
Suffix definable
valid: Provide MatchRegex (later maybe moregeneral) to check if for a given
    file which is currently processed, this stream shall be used.
mode: Append, Write with overwrite, write without overwrite,
    write with error if would overwrite

open: function opening the stream and providing an element to access it
close: function for closing the access element
convert: function to convert the input to write into a valid format
    for the stream (e.g. encode to bytes)
write: function (access element, *args, **kwargs) to write *args, **kwargs
    into the stream


---OpenStreamDefinition
Creates a simple python file stream like 'open'
Can be parsed by 'open(...)'

---GzipStreamDefinition
Creates a gzip python stream
Can be parsed by 'gzip(...)'


--StreamContext
Context Manager to manage 0-N stream definitions. On 'next(arg)' informs all
its StreamDefinitions that the following writes are for another source file.
Is a ContextManager such that it can be used with 'with' and closes streams
if required on __exit__ (e.g. if exit because of exception).
If no stream is given, all writes to it do nothing, if multiple are given,
then the writes are given to all streams (e.g. one stream which check
if the source problem starts with 'verification_' and stores the data in
a verification set and one other stream for a training set)


SamplingBridges
As some parameters are now encoded in the StreamDefinitions (and thus can be
set per stream), those parameters have been removed. Instead, the
SamplingBridges have now a list of StreamContext to which the samples
are feed and which forwards the storing to its streams (the loading is
still done in the ancient way, somewhen the StreamDefinitions could also
include the definitions on how to read data sets made by them. The current
way is trying all techniques until one succeeds).


fast-training.py & fast-sampling.py are adapted to those changes. Mainly
this means, the streams have to be defined in the command line argument
and are parsed by the framework. As the streams contain now some parameters
previously encoded in argparser, those arguemtns have disappeared from
fast-training and fast-sampling argparsers.

fast-training allows now also to split a part of the training data off and
add it to the verification data (previously the verification data had to
be a different data set)