Refactoring Sampling Bridge & Expanding Parser for nested List & Maps
The Sampling Bridges are now a custom subpackage (of the bridge module)
and not anymore a simple module in the bridge package. The __init__.py
loads the same functionality which previously was available in
sampling_bridge.py, thus, no changes should be noticable for existing
scripts.

the TreeParser was improved. Now any nesting of lists via [] and maps
via {k:v} can be parsed. If given as argument a nested list/map thing,
then the nesting is automatically parsed, but the final values in the
map resp. list will be parsed like the field in the object requires (
if the ClassArgument describes the field to be a string, all items in
the final containers of the nesting are parsed as string